requirements.txt
torch>=1.9.0
tqdm>=4.57.0
boto3>=1.17.15
CMRESHandler>=1.0.0
jsonschema>=3.2.0
Deprecated>=1.2.11
opencv-python>=4.5.1
scipy>=1.6.1
matplotlib>=3.3.4
psutil>=5.8.0
tensorboard>=2.4.1
setuptools>=21.0.0
coverage~=5.3.1
torchvision>=0.10.0
sphinx~=4.0.2
sphinx-rtd-theme
torchmetrics>=0.5.0
pillow>=6.2.0
hydra-core
omegaconf
onnxruntime==1.8.1
onnx>=1.10.1
pip-tools>=6.4.0
tensorflow
pyparsing==2.4.5